db.py and fastapi.py moved to own route directories
