+++ obi_one/__init__.py
@@ -5,10 +5,10 @@ from obi_one.core.single import SingleCoordinateMixin
 from obi_one.core.block import Block
 from obi_one.core.scan import GridScan, CoupledScan
 from obi_one.core.serialization import deserialize_obi_object_from_json_file, deserialize_obi_object_from_json_data
-from obi_one.core.fastapi import activate_fastapi_app
-from obi_one.core.db_old import database, circuits, close_db
-from obi_one.core.db_old import CircuitEntity, CircuitCollectionEntity, circuit_collections, circuits
-from obi_one.core.db import init_db, entitysdk_classes
+
+from obi_one.fastapi.fastapi import activate_fastapi_app
+
+from obi_one.db.db import init_db, entitysdk_classes
 for cls in entitysdk_classes:
     globals()[cls.__name__] = cls
