+++ obi_one/modeling/morphology_metrics/morphology_metrics.py
@@ -1,10 +1,9 @@
 from obi_one.core.form import Form
 from obi_one.core.block import Block
 from obi_one.core.single import SingleCoordinateMixin
-# from obi_one.core.db import SaveCircuitEntity, SaveCircuitCollectionEntity
 from obi_one.core.path import NamedPath
 
-from obi_one.core.db import *
+from obi_one.db.db import *
 
 class MorphologyMetricsForm(Form):
     """
