+++ obi_one/modeling/morphology_metrics/morphology_metrics_example.py
@@ -6,7 +6,7 @@ from obi_one.core.block import Block
 from obi_one.core.single import SingleCoordinateMixin
 from obi_one.core.path import NamedPath
 from obi_one.core.base import OBIBaseModel
-from obi_one.core.db import *
+from obi_one.db.db import *
 
 class MorphologyMetricsExampleForm(Form):
     """
